refactor algolia so it can be called from other Go pkgs
